rope: move Peer to sparate file
