+++ src/rope/mod.rs
@@ -1,12 +1,11 @@
 use boringtun::crypto::{X25519PublicKey, X25519SecretKey};
-use boringtun::noise::errors::WireGuardError;
-use boringtun::noise::{Tunn, TunnResult};
+
+use boringtun::noise::TunnResult;
 use log::*;
 use parking_lot::{Mutex, RwLock};
 use rpv6::{BoxedPacket, Packet};
 use std::collections::HashMap;
 use std::io;
-use std::ops::Deref;
 use std::sync::{Arc, Weak};
 use tokio::sync::mpsc;
 
@@ -22,157 +21,11 @@ pub mod exaddr;
 pub type ExternalAddr = exaddr::ExternalAddr;
 mod msg;
 
+mod peer;
+use peer::*;
+pub type Peer = peer::Peer;
+
 pub type Msg = msg::Msg;
-
-pub struct Peer {
-    id: u128,
-    static_public_key: Arc<X25519PublicKey>,
-    txs: RwLock<Vec<Arc<dyn Tx>>>,
-    wgtunn: Arc<Tunn>,
-    current_tx: Mutex<Option<Weak<dyn Tx>>>,
-}
-
-pub enum EncryptedSendError {
-    IOE(io::Error),
-    WireGuard(WireGuardError),
-}
-
-impl Peer {
-    /// Initialise a new peer object in Arc.
-    /// ## Safety
-    /// It's safe to fill any of arguments with MaybeUninit and set them later.
-    /// Make sure all of them are properly set before calling any method.
-    pub fn new_arc(id: u128, static_public_key: Arc<X25519PublicKey>, wgtunn: Arc<Tunn>) -> Arc<Self> {
-        Arc::new(Self::new(id, static_public_key, wgtunn))
-    }
-
-    /// Initialise a new peer object.
-    /// ## Safety
-    /// It's safe to fill any of arguments with MaybeUninit and set them later.
-    /// Make sure all of them are properly set before calling any method.
-    pub fn new(id: u128, static_public_key: Arc<X25519PublicKey>, wgtunn: Arc<Tunn>) -> Self {
-        Self {
-            id,
-            static_public_key,
-            txs: RwLock::new(Vec::new()),
-            wgtunn,
-            current_tx: Mutex::new(None),
-        }
-    }
-
-    /// Scan all txs and remove the ones are removable.
-    /// This function will hold read-write lock on .txs.
-    pub fn gc_tx(&self) {
-        let mut txs = self.txs.write();
-        txs.retain(|v| !v.is_removable());
-    }
-
-    /// Scan all txs and choose one with largest availability.
-    /// This function will hold lock on .current_tx and read-write lock on .txs.
-    fn select_tx_slow(&self) -> Option<Arc<dyn Tx>> {
-        self.gc_tx();
-        let txs = self.txs.read();
-        let best_tx = txs.iter().fold(None, |prev: Option<&Arc<dyn Tx>>, next| match prev {
-            Some(prev_tx) => if next.get_availability() > prev_tx.get_availability() { Some(next) } else { Some(prev_tx) },
-            None => Some(next)
-        }).cloned();
-        *self.current_tx.lock() = match best_tx.clone() {
-            Some(tx) => Some(Arc::downgrade(&tx)),
-            None => None,
-        };
-        debug!("Choose {:?} for peer {}", best_tx, self.id);
-        best_tx
-    }
-
-    fn select_tx(&self, prefer_address: Option<&ExternalAddr>) -> Option<Arc<dyn Tx>> {
-        if let Some(addr) = prefer_address {
-            if let Some(tx) = self.find_tx_of_addr(addr.clone()) {
-                return Some(tx)
-            }
-        }
-        let current_tx = self.current_tx.lock();
-        if let Some(tx_weakref) = current_tx.deref() {
-            if let Some(tx) = tx_weakref.upgrade() {
-                if tx.get_availability() <= 0.6 {
-                    std::mem::drop(current_tx); // avoid deadlock
-                    self.select_tx_slow()
-                } else {
-                    Some(tx)
-                }
-            } else {
-                std::mem::drop(current_tx); // avoid deadlock
-                self.select_tx_slow()
-            }
-        } else {
-            std::mem::drop(current_tx); // avoid deadlock
-            self.select_tx_slow()
-        }
-    }
-
-
-    /// Encrypt src, write to dst, and send.
-    /// The size of dst should be src.len() + 32 and should not be less than 148 bytes.
-    async fn send_encrypted<'a>(&self, src: &[u8], dst: &'a mut [u8]) -> Result<usize, EncryptedSendError> {
-        let mut data = src;
-        let mut tx_bytes = 0usize;
-        loop {
-            match self.wgtunn.encapsulate(data, dst) {
-                TunnResult::Done => break Ok(tx_bytes),
-                TunnResult::WriteToNetwork(buf) => {
-                    match self.send(buf, None).await {
-                        Ok(size) => tx_bytes += size,
-                        Err(e) => break Err(EncryptedSendError::IOE(e)),
-                    }
-                },
-                TunnResult::WriteToTunnelV4(_, _) => unreachable!(),
-                TunnResult::WriteToTunnelV6(_, _) => unreachable!(),
-                TunnResult::Err(e) => break Err(EncryptedSendError::WireGuard(e)),
-            };
-            data = &[];
-        }
-    }
-
-    /// Send cleartext though any tx.
-    /// Return WouldBlock if no tx avaliable.
-    async fn send(&self, data: &[u8], prefer_address: Option<&ExternalAddr>) -> io::Result<usize> {
-        if let Some(tx) = self.select_tx(prefer_address) {
-            tx.send_to(data).await
-        } else {
-            Err(io::ErrorKind::WouldBlock.into())
-        }
-    }
-
-    pub fn add_tx(&self, tx: Box<dyn Tx>) {
-        let mut txs = self.txs.write();
-        txs.push(Arc::from(tx));
-    }
-
-    pub fn clear_tx(&self) {
-        let mut txs = self.txs.write();
-        txs.clear();
-        txs.shrink_to(0);
-    }
-
-    pub fn get_id(&self) -> u128 {
-        self.id
-    }
-
-    pub fn find_tx_of_addr(&self, exaddr: ExternalAddr) -> Option<Arc<dyn Tx>> {
-        let txs = self.txs.read();
-        let mut result = None;
-        for tx in txs.iter() {
-            if tx.is_match_addr(exaddr.clone()) {
-                result =  Some(tx.clone());
-                break;
-            }
-        }
-        result
-    }
-
-    pub fn get_public_key(&self) -> Arc<X25519PublicKey> {
-        self.static_public_key.clone()
-    }
-}
 
 async fn router_routing_rx_thread_body(
     router_ref: Weak<Router>,
@@ -371,7 +224,7 @@ impl Router {
     pub fn find_peer_by_public_key(&self, key: &X25519PublicKey) -> Option<Arc<Peer>> {
         let peers = self.peers.read();
         for peer in peers.iter() {
-            if let Ok(_) = peer.static_public_key.constant_time_is_equal(key) {
+            if let Ok(_) = peer.get_public_key().constant_time_is_equal(key) {
                 return Some(peer.clone());
             }
         }
@@ -381,7 +234,7 @@ impl Router {
     pub fn find_peer_by_id(&self, id: u128) -> Option<Arc<Peer>> {
         let peers = self.peers.read();
         for peer in peers.iter() {
-            if peer.id == id {
+            if peer.get_id() == id {
                 return Some(peer.clone());
             }
         }
